Add temp version of unified physical_data for partition API
