+++ modin/core/dataframe/pandas/dataframe/dataframe.py
@@ -2528,7 +2528,7 @@ class PandasDataframe(object, metaclass=LoggerMetaClass):
         # unwrap list returned by `copartition`.
         right_parts = right_parts[0]
         new_frame = self._partition_mgr_cls.binary_operation(
-            left_parts, lambda l, r: op(l, r), right_parts, axis=1
+            left_parts, lambda l, r: op(l, r), right_parts
         )
 
         return self.__constructor__(
