+++ modin/core/execution/dask/implementations/pandas_on_dask/partitioning/partition_manager.py
@@ -13,12 +13,8 @@
 
 """Module houses class that implements ``PandasDataframePartitionManager``."""
 
-import numpy as np
-import pandas
-
 from modin.core.dataframe.pandas.partitioning.partition_manager import (
     PandasDataframePartitionManager,
-    wait_computations_if_benchmark_mode,
 )
 from modin.core.execution.dask.common.engine_wrapper import DaskWrapper
 from .virtual_partition import (
@@ -52,46 +48,3 @@ class PandasOnDaskDataframePartitionManager(PandasDataframePartitionManager):
             The objects wrapped by `partitions`.
         """
         return DaskWrapper.materialize([partition.future for partition in partitions])
-
-    @classmethod
-    @wait_computations_if_benchmark_mode
-    def binary_operation(cls, left, func, right, axis=1):
-        """
-        Apply a function that requires partitions of two ``PandasOnRayDataframe`` objects.
-
-        Parameters
-        ----------
-        left : np.ndarray
-            The partitions of left ``PandasOnRayDataframe``.
-        func : callable
-            The function to apply.
-        right : np.ndarray
-            The partitions of right ``PandasOnRayDataframe``.
-        axis : {0, 1}, default: 1
-            The axis to apply the function over (0 - rows, 1 - columns).
-
-        Returns
-        -------
-        np.ndarray
-            A NumPy array with new partitions.
-        """
-        [part.drain_call_queue() for row in right for part in row]
-
-        def op_with_empty_check(x, y, *args, **kwargs):
-            y = pandas.DataFrame(index=x.index, columns=x.columns) if y is None else y
-
-            return func(x, y, *args, **kwargs)
-
-        op_with_empty_check = cls.preprocess_func(op_with_empty_check)
-        return np.array(
-            [
-                [
-                    part.apply(
-                        op_with_empty_check,
-                        right[row_idx][col_idx].oid if len(right) else None,
-                    )
-                    for col_idx, part in enumerate(left[row_idx])
-                ]
-                for row_idx in range(len(left))
-            ]
-        )
